+++ test/services/supplementary-billing/fetch-invoice-account-numbers.service.test.js
@@ -12,9 +12,9 @@ const DatabaseHelper = require('../../support/helpers/database.helper.js')
 const InvoiceAccountHelper = require('../../support/helpers/crm-v2/invoice-account.helper.js')
 
 // Thing under test
-const FetchInvoiceAccountsService = require('../../../app/services/supplementary-billing/fetch-invoice-accounts.service.js')
+const FetchInvoiceAccountNumbersService = require('../../../app/services/supplementary-billing/fetch-invoice-account-numbers.service.js')
 
-describe('Fetch Invoice Accounts service', () => {
+describe('Fetch Invoice Account Numbers service', () => {
   beforeEach(async () => {
     await DatabaseHelper.clean()
   })
@@ -32,13 +32,13 @@ describe('Fetch Invoice Accounts service', () => {
       ])
 
       expectedResult = [
-        invoiceAccounts[0],
-        invoiceAccounts[1]
+        { invoiceAccountId: invoiceAccounts[0].invoiceAccountId, invoiceAccountNumber: invoiceAccounts[0].invoiceAccountNumber },
+        { invoiceAccountId: invoiceAccounts[1].invoiceAccountId, invoiceAccountNumber: invoiceAccounts[1].invoiceAccountNumber }
       ]
     })
 
     it('fetches the invoice accounts that the charge versions link to', async () => {
-      const result = await FetchInvoiceAccountsService.go([
+      const result = await FetchInvoiceAccountNumbersService.go([
         { invoiceAccountId: invoiceAccounts[0].invoiceAccountId },
         { invoiceAccountId: invoiceAccounts[1].invoiceAccountId }
       ])
